+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 1.0.0
 -----
 
+- #84 Add date range filter toggle and refactor date handling logic
 - #76 Add script to export SENAITE analysis requests by date range
 - #83 Add Tamanu ID Field
 - #75 Add date range search function in samples listing